fix sizes calc

+++ src/FMIndex.hpp
@@ -15,5 +15,5 @@ class FMIndex {
     std::vector<size_t> doc_locate(const std::string &pattern);
 
     // Retornar tamaño en megabytes
-    size_t get_size_mb() { return sdsl::size_in_mega_bytes(fm_index); };
+    double get_size_mb() { return sdsl::size_in_mega_bytes(fm_index); };
 };
